Move the flyover lookup wholesale into the XHTML parsing hook
Relocate getTextDisplayFlyover from utils/extensions to hooks/useParseXhtml
unchanged, instead of splitting it into a data half and a parsing wrapper.
The function has a single consumer (useRenderingExtensions), and the
traversal only needs isSpecificItemControl and getXHtmlString, which stay
exported from utils/extensions. This keeps html-react-parser out of the
core extension reader with a smaller diff and drops the
TextDisplayFlyoverSource interface introduced by the earlier split.

Move the five pre-existing flyover tests verbatim to useParseXhtml.test.tsx,
keep the fall-through test, and drop the wrapper tests that duplicated
traversal coverage.
