+++ packages/smart-forms-renderer/src/test/extension.test.tsx
@@ -32,7 +32,6 @@ import {
   getRegexValidation,
   getRequiredFeedback,
   getShortText,
-  getTextDisplayFlyoverSource,
   getTextDisplayInstructions,
   getTextDisplayLower,
   getTextDisplayPrompt,
@@ -1122,183 +1121,6 @@ describe('getTextDisplayInstructions', () => {
       ]
     };
     expect(getTextDisplayInstructions(qItem)).toBe('');
-  });
-});
-
-describe('getTextDisplayFlyoverSource', () => {
-  const ITEM_CONTROL_URL = 'http://hl7.org/fhir/StructureDefinition/questionnaire-itemControl';
-
-  it('returns the raw xHtmlString when getXHtmlString returns a value', () => {
-    const qItem: QuestionnaireItem = {
-      linkId: 'q1',
-      type: 'group',
-      item: [
-        {
-          linkId: 'q1-child-flyover',
-          type: 'display',
-          text: 'flyover fallback text',
-          _text: {
-            extension: [
-              {
-                url: 'http://hl7.org/fhir/StructureDefinition/rendering-xhtml',
-                valueString: '<div xmlns="http://www.w3.org/1999/xhtml">Flyover from XHTML</div>'
-              }
-            ]
-          },
-          extension: [
-            {
-              url: ITEM_CONTROL_URL,
-              valueCodeableConcept: {
-                coding: [
-                  {
-                    system: 'http://hl7.org/fhir/questionnaire-item-control',
-                    code: 'flyover'
-                  }
-                ]
-              }
-            }
-          ]
-        }
-      ]
-    };
-
-    // The extension reader stays DOM-free, so it hands back the unparsed markup
-    expect(getTextDisplayFlyoverSource(qItem)).toEqual({
-      xHtmlString: '<div xmlns="http://www.w3.org/1999/xhtml">Flyover from XHTML</div>',
-      text: ''
-    });
-  });
-
-  it('returns childItem.text if no XHTML string', () => {
-    const qItem: QuestionnaireItem = {
-      linkId: 'q1',
-      type: 'group',
-      item: [
-        {
-          linkId: 'q1-child-flyover',
-          type: 'display',
-          text: 'flyover plain text',
-          extension: [
-            {
-              url: ITEM_CONTROL_URL,
-              valueCodeableConcept: {
-                coding: [
-                  {
-                    system: 'http://hl7.org/fhir/questionnaire-item-control',
-                    code: 'flyover'
-                  }
-                ]
-              }
-            }
-          ]
-        }
-      ]
-    };
-
-    expect(getTextDisplayFlyoverSource(qItem)).toEqual({
-      xHtmlString: null,
-      text: 'flyover plain text'
-    });
-  });
-
-  it('returns empty text if there is no child item', () => {
-    const qItem: QuestionnaireItem = {
-      linkId: 'q1',
-      type: 'group'
-    };
-    expect(getTextDisplayFlyoverSource(qItem)).toEqual({ xHtmlString: null, text: '' });
-  });
-
-  it('handles item.text missing gracefully', () => {
-    const qItem: QuestionnaireItem = {
-      linkId: 'q1',
-      type: 'group',
-      item: [
-        {
-          linkId: 'q1-child-flyover',
-          type: 'display',
-          extension: [
-            {
-              url: ITEM_CONTROL_URL,
-              valueCodeableConcept: {
-                coding: [
-                  {
-                    system: 'http://hl7.org/fhir/questionnaire-item-control',
-                    code: 'flyover'
-                  }
-                ]
-              }
-            }
-          ]
-        }
-      ]
-    };
-    expect(getTextDisplayFlyoverSource(qItem)).toEqual({ xHtmlString: null, text: '' });
-  });
-
-  it('returns empty text if display child type is not "display"', () => {
-    const qItem: QuestionnaireItem = {
-      linkId: 'q1',
-      type: 'group',
-      item: [
-        {
-          linkId: 'q1-child-instructions',
-          type: 'group',
-          text: 'flyover plain text',
-          extension: [
-            {
-              url: ITEM_CONTROL_URL,
-              valueCodeableConcept: {
-                coding: [
-                  {
-                    system: 'http://hl7.org/fhir/questionnaire-item-control',
-                    code: 'flyover'
-                  }
-                ]
-              }
-            }
-          ]
-        }
-      ]
-    };
-    expect(getTextDisplayFlyoverSource(qItem)).toEqual({ xHtmlString: null, text: '' });
-  });
-
-  it('falls through to a later flyover childItem when the first carries neither XHTML nor text', () => {
-    const flyoverExtension: Extension = {
-      url: ITEM_CONTROL_URL,
-      valueCodeableConcept: {
-        coding: [
-          {
-            system: 'http://hl7.org/fhir/questionnaire-item-control',
-            code: 'flyover'
-          }
-        ]
-      }
-    };
-
-    const qItem: QuestionnaireItem = {
-      linkId: 'q1',
-      type: 'group',
-      item: [
-        {
-          linkId: 'q1-child-flyover-empty',
-          type: 'display',
-          extension: [flyoverExtension]
-        },
-        {
-          linkId: 'q1-child-flyover-text',
-          type: 'display',
-          text: 'second flyover',
-          extension: [flyoverExtension]
-        }
-      ]
-    };
-
-    expect(getTextDisplayFlyoverSource(qItem)).toEqual({
-      xHtmlString: null,
-      text: 'second flyover'
-    });
   });
 });
